Encode the ‘no snake loops’ constraint more efficiently
By introducing an array of snake coordinates, we
can more efficiently encode the constraint that
there is only one region of ‘snake’, i.e. there
are no separate loops that circumvent the
restriction that there is only one head/tail in
the grid.

We need to do some busywork to make sure the snake
and grid correspond with each other.