Skip past sets of numbers that are supersets of those already found
For example, if a 1 in the top-left corner uniquely defines a solution
then combining it with any other numbers will also define a unique
solution.

We can terminate the program if we’ve iterated through every combination
of N numbers and none of them have contributed any more minimal sets of
numbers. That would mean we’ve found all starting sets.